Calculate ridgeline data from elevation

Takes elevation data and calculates the data needed for plotting ridgelines by sampling the elevation at regular y-intervals. You will only need this function if you want to create your own custom plots from the ground up.

Usage

calculate_ridgelines(elevation = NULL, n_lines = 30)

Arguments

elevation

A raster object containing elevation data, or NULL to use last elevation

n_lines

Number of ridgelines to calculate

Value

A data frame containing the ridgeline data with columns x, y, elevation, and group

See also